refactor(#7306): the search core moves to engine, so embedded users reach it
The vector, hybrid and full-text search core sat in com.arcadedb.server.vector,
which put it out of reach of embedded ArcadeDB entirely. Nothing about the code
required that: every import is an engine type - Database, RID, LSMVectorIndex,
LSMSparseVectorIndex, FullTextSearch, TypeIndex, JsonSerializer - and not one
line touches com.arcadedb.server, Undertow or an exchange. The placement was
arbitrary, and it was the only thing making these ~1,450 lines server-only.

Embedded already had the primitives: the vectorNeighbors / vectorSparseNeighbors
/ vectorDiscover / vectorRecommend SQL functions and
LSMVectorIndex.findNeighborsFromVector. What it could not reach was the composed
layer above them - RRF and score fusion across legs, graph expansion, leg
weighting and overfetch, and the bounds (MAX_SEEDS, MAX_EXPANSION, MAX_DEPTH,
MAX_LEG_CANDIDATES) that make a request safe. Leaving that in server would have
forced an embedded graph-RAG user to re-derive 690 lines of fusion logic, which
is the same divergence issue #7306 was filed about, one module further out.

com.arcadedb.server.vector -> com.arcadedb.query.search, beside
com.arcadedb.query.sql rather than implying vectors are all that lives there.
HTTP, grpcw and the MCP tools are unchanged apart from the import, and still
share the one implementation and one set of bounds.

No pom.xml changes: engine already carries jvector and Lucene in compile scope,
so the move adds no dependency to any module.

HybridSearchSeedsTest moves with the code - it reaches HybridSearch.MAX_SEEDS by
same-package access. Issue5413DenseVectorServerLatencyIT stays in
com.arcadedb.server.vector: it exercises server latency and never touches the
moved classes.
