+++ engine/src/main/java/com/arcadedb/query/search/VectorSearch.java
@@ -16,7 +16,7 @@
  * SPDX-FileCopyrightText: 2021-present Arcade Data Ltd (info@arcadedata.com)
  * SPDX-License-Identifier: Apache-2.0
  */
-package com.arcadedb.server.vector;
+package com.arcadedb.query.search;
 
 import com.arcadedb.database.Database;
 import com.arcadedb.database.Document;
