+++ grpc-client/src/test/java/com/arcadedb/remote/grpc/Issue7306GrpcVectorSearchIT.java
@@ -30,8 +30,8 @@ import com.arcadedb.server.grpc.HybridSearchResponse;
 import com.arcadedb.server.grpc.SearchHit;
 import com.arcadedb.server.grpc.VectorSearchRequest;
 import com.arcadedb.server.grpc.VectorSearchResponse;
-import com.arcadedb.server.vector.FullTextQuery;
-import com.arcadedb.server.vector.VectorLeg;
+import com.arcadedb.query.search.FullTextQuery;
+import com.arcadedb.query.search.VectorLeg;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
