+++ mcp/src/test/java/com/arcadedb/mcp/Issue7306HttpAndMcpSearchEquivalenceIT.java
@@ -26,8 +26,8 @@ import com.arcadedb.serializer.json.JSONArray;
 import com.arcadedb.serializer.json.JSONObject;
 import com.arcadedb.server.BaseGraphServerTest;
 import com.arcadedb.server.security.ServerSecurityUser;
-import com.arcadedb.server.vector.FullTextQuery;
-import com.arcadedb.server.vector.VectorLeg;
+import com.arcadedb.query.search.FullTextQuery;
+import com.arcadedb.query.search.VectorLeg;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
