+++ mcp/src/test/java/com/arcadedb/mcp/tools/Issue6837ToolInputBoundsFollowUpTest.java
@@ -25,7 +25,7 @@ import com.arcadedb.serializer.json.JSONArray;
 import com.arcadedb.serializer.json.JSONObject;
 import com.arcadedb.server.BaseGraphServerTest;
 import com.arcadedb.server.security.ServerSecurityUser;
-import com.arcadedb.server.vector.VectorLeg;
+import com.arcadedb.query.search.VectorLeg;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
